refactor: Use simpler structure
