+++ gui/app/components/dataloader_creation.py
@@ -5,7 +5,7 @@ from collections.abc import Callable
 import reflex as rx
 from reflex_ag_grid import ag_grid
 
-from ..common import SIDEBAR_OPTIONS, control_buttons, home, select_mode, title
+from .common import SIDEBAR_OPTIONS, control_buttons, home, select_mode, title
 
 
 def checkboxes(row: list[str], state: rx.State) -> rx.Component:
