--- gui/app/pages/__init__.py
@@ -1,1 +1,0 @@
-"""Includes the implementation of pages."""
